+++ threedi_schematisation_editor/utils.py
@@ -2,6 +2,8 @@
 import os
 import shutil
 import sys
+import warnings
+
 from enum import Enum, IntEnum
 from itertools import groupby
 from operator import attrgetter, itemgetter
@@ -746,18 +748,23 @@ def migrate_schematisation_schema(schematisation_filepath, progress_callback=Non
         migration_feedback_msg = f"{e}"
 
     if srid is not None:
+        migration_feedback_msg = ""
         try:
-            schema.upgrade(backup=False, epsg_code_override=srid, progress_func=progress_callback)
+            with warnings.catch_warnings(record=True) as w:
+                warnings.simplefilter("always", UserWarning)
+                schema.upgrade(backup=False, epsg_code_override=srid, progress_func=progress_callback)
+            if w:
+                for warning in w:
+                    migration_feedback_msg += f'{warning._category_name}: {warning.message}\n'
             shutil.rmtree(os.path.dirname(backup_filepath))
             migration_succeed = True
-            migration_feedback_msg = "Migration succeeded."
         except errors.UpgradeFailedError:
-            migration_feedback_msg = (
+            migration_feedback_msg += (
                 "The schematisation database schema cannot be migrated to the current version. "
                 "Please contact the service desk for assistance."
             )
         except Exception as e:
-            migration_feedback_msg = f"{e}"
+            migration_feedback_msg += f"{e}"
 
     return migration_succeed, migration_feedback_msg